requirement.txt
beautifulsoup4==4.4.0
Django==1.8.5
django-admin-bootstrapped==2.5.6
django-app-namespace-template-loader==0.3
django-appconf==1.0.1
django-blog-zinnia==0.16.dev0
django-ckeditor==5.0.2
django-compressor==1.5
django-contrib-comments==1.6.1
django-mptt==0.7.4
django-tagging==0.4
django-taggit==0.17.3
django-xmlrpc==0.1.5
Jinja2==2.8
mots-vides==2015.5.11
Pillow==3.0.0
sorl-thumbnail==12.3
zinnia-theme-bootstrap==0.4
zinnia-wysiwyg-ckeditor==1.3
